Fix unhandled exceptions on wrong audit calls from Rails/RAW SQL
We also added an exception handling here. Problem here is that, when trigger executed and temporary table does not exists (for example we forgot to send :bulk_audit commands) then there is a issue raises. In this case, if anything goes wrong and transaction is not found, whole object will return immediately without breaking anything.